:wrench: Update Plugins' copy destination path
This commit updates the output path of each directory to reflect the
changed build path in the previous commit
(GameHelper/bin/Debug/net5.0/win-64).

I am not aware of a way to add it to all RunTimeIdentifier paths, so I
hardcoded it to win-x64 (which is the only one at this moment).